+++ worldeditadditions_commands/commands/selectors/init.lua
@@ -9,18 +9,23 @@
 local wea_cmdpath = worldeditadditions_commands.modpath .. "/commands/selectors/"
 local weac = worldeditadditions_core
 
-dofile(wea_cmdpath.."srel.lua")
 dofile(wea_cmdpath.."scentre.lua")
 dofile(wea_cmdpath.."scloud.lua")
-dofile(wea_cmdpath.."scol.lua")
-dofile(wea_cmdpath.."scube.lua")
-dofile(wea_cmdpath.."sfactor.lua")
+dofile(wea_cmdpath.."sgrow.lua")
 dofile(wea_cmdpath.."smake.lua")
 dofile(wea_cmdpath.."spop.lua")
 dofile(wea_cmdpath.."spush.lua")
-dofile(wea_cmdpath.."srect.lua")
+dofile(wea_cmdpath.."srel.lua")
 dofile(wea_cmdpath.."sshift.lua")
+dofile(wea_cmdpath.."sshrink.lua")
 dofile(wea_cmdpath.."sstack.lua")
+
+--- DEPRECATED ---
+dofile(wea_cmdpath.."scol.lua")
+dofile(wea_cmdpath.."scube.lua")
+dofile(wea_cmdpath.."srect.lua")
+dofile(wea_cmdpath.."sfactor.lua")
+--- END DEPRECATED ---
 
 dofile(wea_cmdpath.."unmark.lua")
 dofile(wea_cmdpath.."mark.lua")
@@ -29,6 +34,13 @@ dofile(wea_cmdpath.."reset.lua")
 
 -- Aliases
 weac.register_alias("sfac", "sfactor")
+
+weac.register_alias("sgrow", "expand", true) -- true = override target
+weac.register_alias("sgrow", "outset", true) -- true = override target
+weac.register_alias("sshrink", "contract", true) -- true = override target
+weac.register_alias("sshrink", "inset", true) -- true = override target
+
+weac.register_alias("sshift", "shift", true) -- true = override target
 
 weac.register_alias("1", "pos1", true) -- true = override target
 weac.register_alias("2", "pos2", true) -- true = override target