+++ worldeditadditions_commands/commands/selectors/scentre.lua
@@ -16,19 +16,17 @@ worldeditadditions_core.register_command("scentre", {
 	end,
 	func = function(name)
 		local mean = Vector3.mean(
-			Vector3.clone(worldedit.pos1[name]),
-			Vector3.clone(worldedit.pos2[name])
+			Vector3.clone(wea_c.pos.get(name, 1)),
+			Vector3.clone(wea_c.pos.get(name, 2))
 		)
 		local pos1, pos2 = Vector3.clone(mean), Vector3.clone(mean)
 		
 		pos1 = pos1:floor()
 		pos2 = pos2:ceil()
 		
-		worldedit.pos1[name], worldedit.pos2[name] = pos1, pos2
-		worldedit.mark_pos1(name)
-		worldedit.mark_pos2(name)
+		wea_c.pos.set_all(name, {pos1, pos2})
 		
-		return true, "position 1 set to "..pos1..", position 2 set to "..pos2
+		return true, "Position 1 to "..pos1..", Position 2 to "..pos2
 	end,
 })